+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Semantic versioning
 Version latest (0.5.0?, 2019-09-24?)
 
 - ...
+- enh: use `isitfit-tags-suggestBasic-...csv` instead of `...suggest...` to highlight that an advanced suggestion is available
 
 
 Version 0.8.0 (2019-10-04)